Add OM Prefix types
