+++ sbol3/__init__.py
@@ -30,5 +30,6 @@ from .interface import Interface
 from .implementation import Implementation
 from .attachment import Attachment
 from .provenance import Activity, Agent, Association, Plan, Usage
+from .om_prefix import SIPrefix
 from .measurement import Measure, PrefixedUnit, SingularUnit
 from .compound_unit import UnitMultiplication
