+++ sbol3/constants.py
@@ -240,6 +240,7 @@ OM_NS = 'http://www.ontology-of-units-of-measure.org/resource/om-2/'
 
 OM_ALTERNATIVE_LABELS = OM_NS + 'alternativeLabels'
 OM_ALTERNATIVE_SYMBOLS = OM_NS + 'alternativeSymbols'
+OM_BINARY_PREFIX = OM_NS + 'BinaryPrefix'
 OM_COMMENT = OM_NS + 'comment'
 OM_HAS_BASE = OM_NS + 'hasBase'
 OM_HAS_DENOMINATOR = OM_NS + 'hasDenominator'
@@ -256,7 +257,8 @@ OM_LONG_COMMENT = OM_NS + 'longComment'
 OM_MEASURE = OM_NS + 'Measure'
 OM_PREFIXED_UNIT = OM_NS + 'PrefixedUnit'
 OM_SINGULAR_UNIT = OM_NS + 'SingularUnit'
+OM_SI_PREFIX = OM_NS + 'SIPrefix'
 OM_SYMBOL = OM_NS + 'symbol'
-OM_UNIT_MULTIPLICATION = OM_NS + 'UnitMultiplication'
 OM_UNIT_DIVISION = OM_NS + 'UnitDivision'
 OM_UNIT_EXPONENTIATION = OM_NS + 'UnitExponentiation'
+OM_UNIT_MULTIPLICATION = OM_NS + 'UnitMultiplication'
