+++ sbol3/measurement.py
@@ -3,6 +3,7 @@ import math
 from typing import Union
 
 from . import *
+from .om_prefix import Prefix
 
 
 class Unit(TopLevel, abc.ABC):
@@ -24,29 +25,6 @@ class Unit(TopLevel, abc.ABC):
                                                0, math.inf)
         self.comment = TextProperty(self, OM_COMMENT, 0, 1)
         self.long_comment = TextProperty(self, OM_LONG_COMMENT, 0, 1)
-
-
-class Prefix(TopLevel, abc.ABC):
-    """om:Prefix is an abstract base class.
-
-    See Appendix A Section A.2 of the SBOL 3.0 specificiation.
-    """
-
-    def __init__(self, name: str, symbol: str, label: str,
-                 factor: float, type_uri: str) -> None:
-        super().__init__(name, type_uri)
-        self.symbol = TextProperty(self, OM_SYMBOL, 1, 1,
-                                   initial_value=symbol)
-        self.alternative_symbols = TextProperty(self, OM_ALTERNATIVE_SYMBOLS,
-                                                0, math.inf)
-        self.label = TextProperty(self, OM_LABEL, 1, 1,
-                                  initial_value=label)
-        self.alternative_labels = TextProperty(self, OM_ALTERNATIVE_LABELS,
-                                               0, math.inf)
-        self.comment = TextProperty(self, OM_COMMENT, 0, 1)
-        self.long_comment = TextProperty(self, OM_LONG_COMMENT, 0, 1)
-        self.factor = FloatProperty(self, OM_HAS_FACTOR, 1, 1,
-                                    initial_value=factor)
 
 
 class Measure(Identified):
